colorScheme -> viewScheme renaming
